Add Dockerfile to build and package asterctl & sysinfo
Add a Dockerfile to build the asterctl and sysinfo tools. This uses a distroless container as a base, and therefore contains only minimal dependencies - i.e. the built binaries and a handful of dynamic libraries that they depend upon. Using a distroless base keeps the overall size low (~24MB) and reduces noise from vulnerability scanners.

The default configuration, fonts, and CPU/Memory scripts are also copied in to the container image.

Ideally, asterctl and sysinfo would be put in to separate containers to allow someone to just use the tools they want, but for now they're in the same image.